moved all Navigator files to navigator folder
