+++ app/component/itinerary/Navigator/navigator.scss
