+++ sass/_main.scss
@@ -63,7 +63,7 @@ $body-font-weight: $font-weight-medium;
 @import '../app/component/itinerary/settings-panel';
 @import '../app/component/itinerary/mobile-ticket-purchase-information';
 @import 'base/button';
-@import '../app/component/itinerary/navigator';
+@import '../app/component/itinerary/Navigator/navigator';
 @import '../app/component/itinerary/NavigatorIntro/navigator-intro';
 
 /* Modal */
